ToolsUI de-clutter. Split off more OpPanels; visibility of splash.
