+++ ui/src/main/java/ucar/nc2/ui/op/DateFormatMark.java
@@ -3,7 +3,7 @@
  * See LICENSE for license information.
  */
 
-package ucar.nc2.ui.opp;
+package ucar.nc2.ui.op;
 
 import ucar.nc2.units.DateFormatter;
 import ucar.nc2.units.DateFromString;
