+++ ui/src/main/java/ucar/nc2/ui/op/ReportOpPanel.java
@@ -3,21 +3,19 @@
  * See LICENSE for license information.
  */
 
-package ucar.nc2.ui.opp;
+package ucar.nc2.ui.op;
 
 import ucar.nc2.ui.OpPanel;
 import ucar.nc2.ui.ReportPanel;
 import ucar.nc2.ui.widget.BAMutil;
 import ucar.nc2.ui.widget.TextHistoryPane;
 import ucar.util.prefs.PreferencesExt;
-import ucar.nc2.ui.widget.BAMutil;
 
 import java.awt.BorderLayout;
 import java.awt.event.ActionEvent;
 import java.io.IOException;
 import java.io.PrintWriter;
 import java.io.StringWriter;
-import java.util.Formatter;
 import javax.swing.AbstractAction;
 import javax.swing.JComboBox;
 import javax.swing.JOptionPane;
@@ -27,10 +25,10 @@ import javax.swing.JOptionPane;
  */
 public class ReportOpPanel extends OpPanel {
     private ReportPanel reportPanel;
-    boolean useIndex = true;
-    boolean eachFile = false;
-    boolean extra = false;
-    JComboBox reports;
+    private boolean useIndex = true;
+    private boolean eachFile = false;
+    private boolean extra = false;
+    private JComboBox reports;
 
 /**
  *
